fix: preserve claim deadlines and split lease recover tests
Detach ClaimNext cancellation without stripping caller deadlines, and move
ambiguous-claim recovery contracts into a sub-300-line test file.

Generated-By: looper 0.11.0-beta.1 (runner=fixer, agent=grok-build)
